JabBot/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
chromedriver==2.24.1
discord.py==1.4.1
emoji==0.6.0
idna==2.10
lxml==4.5.2
multidict==4.7.6
phantomjs==1.3.0
pokebase==1.3.0
pokedex.py==1.1.2
praw==7.1.0
prawcore==1.5.0
python-dotenv==0.14.0
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
update-checker==0.18.0
urllib3==1.25.10
websocket-client==0.57.0
yarl==1.5.1
